chore: moves PulseSidebar to pulse/sidebar/
