+++ app/src/ui/pulse/sidebar/PulseSidebar.module.scss
