+++ app/src/ui/pulse/sidebar/PulseSidebar.test.tsx
