+++ app/src/ui/pulse/sidebar/PulseSidebar.tsx
