+++ app/src/ui/pulse/sidebar/balance-pill/BalancePill.module.scss
