+++ app/src/ui/pulse/sidebar/balance-pill/BalancePill.tsx
